refactor(j2cl): address CR — extract mention helpers and add test timeout
- Extract duplicate mention helpers (waitForParticipantsJ2cl,
  typeAtMentionTriggerJ2cl, dispatchComposerKeyJ2cl, readMentionStateJ2cl,
  readMentionTriggerLetterJ2cl) from keyboard-shortcuts-parity.spec.ts and
  mention-autocomplete-parity.spec.ts into tests/helpers/mention.ts
- Add test.setTimeout(180_000) and test-user annotation to the newly-enabled
  mention popover ArrowDown/ArrowUp/Enter test to prevent CI flakiness
- Remove hardcoded local path and commit hash from the implementation plan doc
